Move some LogicalPlan util functions into a separate class
